test: extract common code for MT tests
Extract common code from rpma_conn_get_private_data.c
for a future use.

This patch only refactors the code and does not make any functional
modifications.
