+++ tests/multithreaded/conn/CMakeLists.txt
@@ -24,6 +24,7 @@ add_multithreaded(NAME conn BIN rpma_conn_cfg_set_sq_size
 add_multithreaded(NAME conn BIN rpma_conn_cfg_set_timeout
 	SRCS rpma_conn_cfg_set_timeout.c)
 add_multithreaded(NAME conn BIN rpma_conn_get_private_data
-	SRCS rpma_conn_get_private_data.c server_rpma_conn_get_private_data.c)
+	SRCS rpma_conn_get_private_data.c server_rpma_conn_get_private_data.c
+		../common/mtt_connect.c)
 add_multithreaded(NAME conn BIN rpma_conn_req_new
 	SRCS rpma_conn_req_new.c)
